Change server to use Express; fix file directories to work with server
