+++ client/src/html/browse_games.html
@@ -3,10 +3,10 @@
     <head>
         <meta charset="UTF-8" />
         <title>Gamer Port - Recommend Games to Friend in a Minute</title>
-        <script src="browse_games.js" defer></script>
+        <script src="js/browse_games.js" defer></script>
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
         <link rel="stylesheet" href="https://pro.fontawesome.com/releases/v5.10.0/css/all.css" integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p" crossorigin="anonymous"/>
-        <link rel="stylesheet" href="browse_games.css">
+        <link rel="stylesheet" href="css/browse_games.css">
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     </head>
     <body>
@@ -43,10 +43,10 @@
             <div class="navbar-collapse collapse w-100 order-3 dual-collapse2">
                 <ul class="navbar-nav ml-auto">
                     <li class="nav-item">
-                        <a class="nav-link" href="auth_pages/signin.html">Sign In</a>
+                        <a class="nav-link" href="signin.html">Sign In</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="auth_pages/signup.html">Register</a>
+                        <a class="nav-link" href="signup.html">Register</a>
                     </li>
                 </ul>
             </div>
