+++ client/src/html/css/browse_games.css
