+++ client/src/html/css/profile.css
@@ -1,5 +1,5 @@
 body {
-    background-image: url("/ProfileBackground.png");
+    background-image: url("/images/ProfileBackground.png");
   }
 
 .align-to-divider {
