+++ client/src/html/forgot_password.html
@@ -11,7 +11,7 @@
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css"
         integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU" crossorigin="anonymous">
     <!--Custom styles-->
-    <link rel="stylesheet" type="text/css" href="/signin.css">
+    <link rel="stylesheet" type="text/css" href="css/signin.css">
 </head>
 
 <body>
@@ -26,32 +26,32 @@
         <div class="mx-auto order-0">
             <ul class="navbar-nav ml-auto">
                 <li class="nav-item">
-                    <a class="nav-link" href="/profile.html">Profile</a>
+                    <a class="nav-link" href="profile.html">Profile</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/browse_games.html">Browse</a>
+                    <a class="nav-link" href="browse_games.html">Browse</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/recommendations.html">Recommendations</a>
+                    <a class="nav-link" href="recommendations.html">Recommendations</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/games.html">Games</a>
+                    <a class="nav-link" href="games.html">Games</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/wishlist.html">Wishlist</a>
+                    <a class="nav-link" href="wishlist.html">Wishlist</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/messages.html">Messages</a>
+                    <a class="nav-link" href="messages.html">Messages</a>
                 </li>
             </ul>
         </div>
         <div class="navbar-collapse collapse w-100 order-3 dual-collapse2">
             <ul class="navbar-nav ml-auto">
                 <li class="nav-item">
-                    <a class="nav-link" href="/auth_pages/signin.html">Sign In</a>
+                    <a class="nav-link" href="signin.html">Sign In</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/auth_pages/signup.html">Register</a>
+                    <a class="nav-link" href="signup.html">Register</a>
                 </li>
             </ul>
         </div>
@@ -60,37 +60,20 @@
         <div class="row h-100 justify-content-center align-middle">
             <div class="card">
                 <div class="card-header">
-                    <h3>Registration</h3>
+                    <h3>Password Reset</h3>
                 </div>
                 <div class="card-body">
                     <form>
-                        <div class="input-group form-group">
-                            <div class="input-group-prepend">
-                                <span class="input-group-text"><i class="fas fa-envelope"></i></span>
-                            </div>
-                            <input type="text" class="form-control" placeholder="username">
-                        </div>
                         <div class="input-group form-group">
                             <div class="input-group-prepend">
                                 <span class="input-group-text"><i class="fas fa-user"></i></span>
                             </div>
                             <input type="text" class="form-control" placeholder="email">
                         </div>
-                        <div class="input-group form-group">
-                            <div class="input-group-prepend">
-                                <span class="input-group-text"><i class="fas fa-key"></i></span>
-                            </div>
-                            <input type="password" class="form-control" placeholder="password">
-                        </div>
                         <div class="form-group d-flex justify-content-center">
-                            <input type="submit" value="Login" class="btn float-right login_btn">
+                            <input type="submit" value="Reset" class="btn float-right login_btn">
                         </div>
                     </form>
-                </div>
-                <div class="card-footer">
-                    <div class="d-flex justify-content-center links">
-                        Have an account?<a href="/auth_pages/signin.html">Sign In</a>
-                    </div>
                 </div>
             </div>
         </div>
