+++ client/src/html/messages.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
     <!--Custom styles-->
-    <link rel="stylesheet" type="text/css" href="messages.css">
+    <link rel="stylesheet" type="text/css" href="css/messages.css">
   </head>
   <body>
     <nav class="navbar navbar-expand-md navbar-dark bg-dark">
@@ -43,10 +43,10 @@
       <div class="navbar-collapse collapse w-100 order-3 dual-collapse2">
         <ul class="navbar-nav ml-auto">
           <li class="nav-item">
-            <a class="nav-link" href="auth_pages/signin.html">Sign In</a>
+            <a class="nav-link" href="signin.html">Sign In</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="auth_pages/signup.html">Register</a>
+            <a class="nav-link" href="signup.html">Register</a>
           </li>
         </ul>
       </div>
