+++ client/src/html/profile.html
@@ -4,7 +4,7 @@
         <title>Gamer Port - Recommend Games to Friend in a Minute</title>
         
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
-        <link rel="stylesheet" href="profile.css">
+        <link rel="stylesheet" href="css/profile.css">
         <meta charset="UTF-8" />
     </head>
     <body>
@@ -41,10 +41,10 @@
             <div class="navbar-collapse collapse w-100 order-3 dual-collapse2">
               <ul class="navbar-nav ml-auto">
                 <li class="nav-item">
-                  <a class="nav-link" href="auth_pages/signin.html">Sign In</a>
+                  <a class="nav-link" href="signin.html">Sign In</a>
                 </li>
                 <li class="nav-item">
-                  <a class="nav-link" href="auth_pages/signup.html">Register</a>
+                  <a class="nav-link" href="signup.html">Register</a>
                 </li>
               </ul>
             </div>
@@ -52,7 +52,7 @@
         <div class="whitebackground">
             <h1 class="header">Profile</h1>
             <div class="center">
-                <img class ="profilepic" src=/blankprofile.png alt="profile picture">
+                <img class ="profilepic" src="images/blankprofile.png" alt="profile picture">
                 <button class="btn btn-secondary btn-sm buttoncenter ml-4" ID="setprofpic">Set Profile Picture</button>
             </div>
             <hr class="divider mb-4">
@@ -88,12 +88,12 @@
             <hr class="divider mt-4 mb-4">
             <h4 class="header mb-4">Friends List</h4>
             <div class="centerfriendlist">
-                <img class ="friendprofilepic mr-4" src=/blankprofile.png alt="friend profile picture">
+                <img class ="friendprofilepic mr-4" src="images/blankprofile.png" alt="friend profile picture">
                 <label>Friend1Username</label>
                 <button class="btn btn-danger btn-sm buttoncenter ml-4" ID="deletefriend1">Remove Friend</button>
             </div>
             <div class="centerfriendlist mb-5">
-                <img class ="friendprofilepic mr-4" src=/blankprofile.png alt="friend profile picture">
+                <img class ="friendprofilepic mr-4" src="images/blankprofile.png" alt="friend profile picture">
                 <label>Friend2Username</label>
                 <button class="btn btn-danger btn-sm buttoncenter ml-4" ID="deletefriend2">Remove Friend</button>
             </div>
@@ -108,7 +108,7 @@
                 </div>
             </div>
             <div class="centerfriendlist">
-                <img class ="friendprofilepic mr-4" src=/blankprofile.png alt="friend profile picture">
+                <img class ="friendprofilepic mr-4" src="images/blankprofile.png" alt="friend profile picture">
                 <label>Friend3Username</label>
                 <button class="btn btn-success btn-sm buttoncenter ml-4" ID="addfriend3">Add Friend</button>
             </div>
