+++ client/src/html/recommendations.html
@@ -4,10 +4,10 @@
         <meta charset="UTF-8" />
         <title>Gamer Port - Recommend Games to Friend in a Minute</title>
         
-        <script src="recommendations.js" defer></script>
+        <script src="js/recommendations.js" defer></script>
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-        <link rel="stylesheet" href="recommendations.css">
+        <link rel="stylesheet" href="css/recommendations.css">
     </head>
     <body>
         <nav class="navbar navbar-expand-md navbar-dark bg-dark">
@@ -43,10 +43,10 @@
             <div class="navbar-collapse collapse w-100 order-3 dual-collapse2">
             <ul class="navbar-nav ml-auto">
                 <li class="nav-item">
-                <a class="nav-link" href="auth_pages/signin.html">Sign In</a>
+                <a class="nav-link" href="signin.html">Sign In</a>
                 </li>
                 <li class="nav-item">
-                <a class="nav-link" href="auth_pages/signup.html">Register</a>
+                <a class="nav-link" href="signup.html">Register</a>
                 </li>
             </ul>
             </div>
